Update the Workout form.
Lift up the form state to the _layout, add reset to the workoutClose
func, update the useFildArray key name to prevent naming conflict
with the exercise id.
